pyproject.toml
[project]
name = "mcpreviewer"
version = "0.1.0"
description = "MCP Reviewer — PR review bot for MCP capability changes"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "PyJWT[crypto]>=2.8",
    "PyYAML>=6.0",
    "click>=8.1",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pip-audit",
]

[project.scripts]
mcpreviewer = "mcpreviewer.cli.main:analyze_cmd"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcpreviewer*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
